+++ tools/lgtdoc/xml/lgt2md.js
@@ -1,7 +1,7 @@
 /////////////////////////////////////////////////////////////////////////////
 //
 //   XML documenting files to Mardown text files conversion script 
-//   Last updated on February 17, 2016
+//   Last updated on February 19, 2016
 //
 //   This file is part of Logtalk <http://logtalk.org/>  
 //   Copyright 1998-2016 Paulo Moura <pmoura@logtalk.org>
@@ -29,6 +29,9 @@ if (ScriptEngineMajorVersion() < 5 || ScriptEngineMajorVersion() == 5 && ScriptE
 var WshShell = new ActiveXObject("WScript.Shell");
 
 var directory = WshShell.CurrentDirectory;
+
+var index_file = "index.md"
+var index_title = "Documentation index"
 
 var processor = "msxsl";
 // var processor = "xsltproc";
@@ -70,10 +73,18 @@ var entity_xslt = logtalk_user + "\\tools\\lgtdoc\\xml\\logtalk_entity_to_md.xsl
 var index_xslt = logtalk_user + "\\tools\\lgtdoc\\xml\\logtalk_index_to_md.xsl";
 
 var d_arg = "";
+var i_arg = "";
+var t_arg = "";
 var p_arg = "";
 
 if (WScript.Arguments.Named.Exists("d"))
 	d_arg = WScript.Arguments.Named.Item("d");
+
+if (WScript.Arguments.Named.Exists("i"))
+	i_arg = WScript.Arguments.Named.Item("i");
+
+if (WScript.Arguments.Named.Exists("t"))
+	t_arg = WScript.Arguments.Named.Item("t");
 
 if (WScript.Arguments.Named.Exists("p"))
 	p_arg = WScript.Arguments.Named.Item("p");
@@ -86,6 +97,12 @@ if (d_arg != "" && !FSObject.FolderExists(d_arg)) {
 	usage_help();
 } else if (d_arg != "")
 	directory = d_arg;
+
+if (i_arg != "")
+	index_file=i_arg;
+
+if (t_arg != "")
+	index_title=t_arg;
 
 if (p_arg != "" && p_arg != "msxsl" && p_arg != "xsltproc" && p_arg != "xalan" && p_arg != "sabcmd") {
 	WScript.Echo("Error! Unsupported XSLT processor:" + p_arg);
@@ -153,6 +170,13 @@ for (files.moveFirst(); !files.atEnd(); files.moveNext()) {
 
 WScript.Echo("conversion done");
 WScript.Echo("");
+WScript.Echo("generating index file...");
+
+index_file = directory + "\\" + index_file;
+create_index_file();
+
+WScript.Echo("index file generated");
+WScript.Echo("");
 
 WScript.Quit(0);
 
@@ -162,12 +186,67 @@ function usage_help() {
 	WScript.Echo("current directory to Markdown text files");
 	WScript.Echo("");
 	WScript.Echo("Usage:");
-	WScript.Echo("  " + WScript.ScriptName + " [/d:directory] [/p:processor]");
+	WScript.Echo("  " + WScript.ScriptName + " [/d:directory] [/i:index] [/t:title] [/p:processor]");
 	WScript.Echo("  " + WScript.ScriptName + " help");
 	WScript.Echo("");
 	WScript.Echo("Optional arguments:");
 	WScript.Echo("  d - output directory for the generated files (default is " + directory + ")");
+	WScript.Echo("  i - name of the index file (default is " + index_file + ")");
+	WScript.Echo("  t - title to be used in the index file (default is " + index_title + ")");
 	WScript.Echo("  p - XSLT processor (msxsl, xsltproc, xalan, or sabcmd; default is " + processor + ")");
 	WScript.Echo("");
 	WScript.Quit(1);
+}
+
+function create_index_file() {
+
+	var f = FSObject.CreateTextFile(index_file, true);
+
+	f.WriteLine("# " + index_title);
+
+	var files = new Enumerator(FSObject.GetFolder(WshShell.CurrentDirectory).Files);
+
+	if (FSObject.FileExists(WshShell.CurrentDirectory + "\\directory_index.xml")) {
+		f.WriteLine("* [Directory index](directory_index.md)");
+		f.WriteLine("* [Entity index](entity_index.md)");
+		f.WriteLine("* [Predicate index](predicate_index.md)");
+	} else {
+		for (files.moveFirst(); !files.atEnd(); files.moveNext()) {
+			var file = files.item().name;
+			if (FSObject.GetExtensionName(file) == "xml") {
+				var md_file = FSObject.GetBaseName(file) + ".md";
+				WScript.Echo("  indexing " + html_file);
+				var index = FSObject.GetBaseName(file).lastIndexOf("_");
+				var pars = FSObject.GetBaseName(file).slice(index+1);
+				var entity = FSObject.GetBaseName(file).slice(0, index);
+				if (pars == 0)
+					f.WriteLine("* [" + entity + "](" + md_file + ")");
+				else
+					f.WriteLine("* [" + entity + "/" + pars + "](" + md_file + ")");
+			}
+		}
+	}
+
+	var today = new Date();
+	var year  = today.getFullYear();
+	var month = today.getMonth() + 1;
+	if (month < 10)
+        month = "0" + month;
+	var day = today.getDate();
+	if (day < 10)
+        day = "0" + day;
+	strToday = year + "/" + month + "/" + day;
+	var hours = today.getHours();
+	if (hours < 10)
+        hours = "0" + hours;
+	var mins = today.getMinutes();
+	if (mins < 10)
+        mins = "0" + mins;
+	var secs = today.getSeconds();
+	if (secs < 10)
+        secs = "0" + secs;
+	strTime = hours + ":" + mins + ":" + secs;
+	f.WriteLine("Generated on " + strToday + " - " + strTime);
+
+	f.Close();
 }